asdf

+++ Application/Home/Controller/ReimburseController.class.php
@@ -622,13 +622,15 @@ class ReimburseController extends Controller
                 $sest = I('post.est');
                 $sast = I('post.ast');
                 if($est != 0 || $ast != 0 ){
-                    //if($est == $sest && $ast == $sast){
+                    if($est == $sest && $ast == $sast){
                         M('oreimburse')->where("recordid = {$id}")->delete();
                         $rses = M('oreimburse')->addAll($dat);
-                    //}else{
-                    //    $this->error("数据之和不匹配！est=".$est.' ast='.$ast);
-                    //    exit;
-                    //}
+                    }else{
+                        $strsss = $est.'<=>'.$sest.'||'.$ast.'<=>'.$sast;
+                        $this->error("数据之和不匹配！".$strsss);
+
+                        exit;
+                    }
                 }
 
                 $bsmid = I('post.bsmid');
